Added timestamps to UserAccount.  Profile settings page added and accessible via index.html.
