+++ src/main/java/com/hiddengems/hiddengems/models/AuthenticationFilter.java
@@ -20,7 +20,7 @@ public class AuthenticationFilter extends HandlerInterceptorAdapter {
     @Autowired
     AuthenticationController authenticationController;
 
-    private static final List<String> whitelist = Arrays.asList("/login", "/register", "/logout", "/css");
+    private static final List<String> whitelist = Arrays.asList("/login", "/register", "/logout", "/css", "/profile/settings");
 
     private static boolean isWhitelisted(String path) {
         for (String pathRoot : whitelist) {
