+++ src/main/resources/templates/profile/settings.html
@@ -8,7 +8,7 @@
 <header th:replace="fragments :: header"></header>
 
 <div>
-    <center><h1 th:text="${'Create User Profile'}">Pass Title To Model!!!</h1></center>
+    <center><h1 th:text="${title}">Pass Title To Model!!!</h1></center>
 </div>
 
 
@@ -35,25 +35,25 @@
         <div class="form-group">
             <label class="col-form-label" for="displayName">Display name:*</label>
             <input type="text" class="form-control" placeholder="Identify yourself!" id="displayName"
-                   th:field="${user.displayName}" maxlength="25" aria-required="true">
-            <div class="invalid-feedback" th:errors="${user.displayName}"></div>
+                   th:field="${userProfile.displayName}" maxlength="25" aria-required="true">
+            <div class="invalid-feedback" th:errors="${userProfile.displayName}"></div>
 
             <label class="col-form-label" for="emailAddress">Email Address:*</label>
             <input type="text" class="form-control" placeholder="marcopolo@example.com" id="emailAddress"
-                   th:field="${user.emailAddress}" maxlength="100" aria-required="true">
-            <div class="invalid-feedback" th:errors="${user.emailAddress}"></div>
+                   th:field="${userProfile.emailAddress}" maxlength="100" aria-required="true">
+            <div class="invalid-feedback" th:errors="${userProfile.emailAddress}"></div>
 
             <label for="aboutMe">About me:</label>
-            <textarea class="form-control" th:field="${user.bio}" id="aboutMe" rows="5" maxlength="500"></textarea>
-            <div class="invalid-feedback" th:errors="${user.bio}"></div>
+            <textarea class="form-control" th:field="${userProfile.bio}" id="aboutMe" rows="5" maxlength="500"></textarea>
+            <div class="invalid-feedback" th:errors="${userProfile.bio}"></div>
 
             <label class="col-form-label" for="zip">Zip Code:</label>
             <input type="text" class="form-control" placeholder="5-digit Zip Code (USA)" id="zip"
-                   th:field="${user.zipCode}" maxlength="5">
-            <div class="invalid-feedback" th:errors="${user.zipCode}"></div>
+                   th:field="${userProfile.zipCode}" maxlength="5">
+            <div class="invalid-feedback" th:errors="${userProfile.zipCode}"></div>
         </div>
 
-        <button type="submit" class="btn btn-primary" th:text="(${submitBtnText})">Submit</button>
+        <button type="submit" class="btn btn-primary">Save</button>
     </form>
 </div>
 <!--End user profile creation form-->
