+++ src/main/resources/templates/register.html
@@ -28,6 +28,8 @@
 </form>
 
 
+
+
 <footer th:replace="fragments :: footer"></footer>
 </body>
 </html>
